22-08-2022

+++ lib/app/data/model/loja_ligado_produto.dart
@@ -2,7 +2,7 @@ class lojasLigadasProduto {
   int? produtoId;
   int? parceiroId;
   String? parceiroNome;
-  double? preco;
+  int? preco;
   String? dataValidad;
   int? estadoStok;
